# .sisyphus/drafts/tmuxweb-openapi.yaml
openapi: 3.0.3
info:
  title: TmuxWeb API
  description: |
    API for TmuxWeb - a web-based tmux session manager with task tracking,
    conversation logging, and summary generation capabilities.
    
    ## Authentication
    TmuxWeb uses cookie-based session authentication:
    - Login with a token via `POST /api/auth/login`
    - Session cookie is HttpOnly, Max-Age=30 days, SameSite=Lax, Secure in production
    - All authenticated endpoints require a valid session cookie
    
    ## paneKey Format
    The paneKey identifies a specific tmux pane: `"{sessionName}:{windowIndex}:{paneIndex}"`
    
    Example: `"hsk:1:0"` = session "hsk", window 1, pane 0
  version: 1.0.0
  contact:
    name: TmuxWeb
servers:
  - url: http://localhost:8215
    description: Local development server
...
paths:
...
  /api/profiles:
    get:
      tags: [Profiles]
      summary: List profiles for current token
      description: Returns all profiles (workspaces) for the authenticated user
      operationId: listProfiles
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of profiles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileListResponse'
              example:
                profiles:
                  - id: 1
                    profile_key: "default"
                    name: "Default"
                    ctime: 1707400000
                    mtime: 1707400000
                  - id: 2
                    profile_key: "work"
                    name: "Work"
                    ctime: 1707400100
                    mtime: 1707400100
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Profiles]
      summary: Create a new profile
      operationId: createProfile
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProfileRequest'
            example:
              name: "Personal"
              profile_key: "personal"
      responses:
        '201':
          description: Profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
              example:
                id: 3
                profile_key: "personal"
                name: "Personal"
                ctime: 1707400200
                mtime: 1707400200
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/groups:
    get:
      tags: [Groups]
      summary: List session groups
      description: Returns all groups for current profile with session counts
      operationId: listGroups
      security:
        - cookieAuth: []
      parameters:
        - name: profile_key
          in: query
          required: true
          schema:
            type: string
          description: Profile key to filter groups
      responses:
        '200':
          description: List of groups
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupListResponse'
              example:
                groups:
                  - id: 1
                    group_name: "AI Projects"
                    sort_order: 10
                    session_count: 3
                    ctime: 1707400000
                    mtime: 1707400000
                  - id: 2
                    group_name: "DevOps"
                    sort_order: 20
                    session_count: 2
                    ctime: 1707400100
                    mtime: 1707400100
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Groups]
      summary: Create a session group
      operationId: createGroup
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
            example:
              profile_key: "default"
              group_name: "New Project"
              sort_order: 30
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
              example:
                id: 3
                group_name: "New Project"
                sort_order: 30
                ctime: 1707400200
                mtime: 1707400200
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/panes/status:
    get:
      tags: [Panes]
      summary: Get status for pane(s)
      description: Returns pane status (idle/in_progress/done) for specified panes
      operationId: getPaneStatus
      security:
        - cookieAuth: []
      parameters:
        - name: pane_keys
          in: query
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
          description: "Pane keys (format: sessionName:windowIndex:paneIndex)"
          example: ["hsk:1:0", "hsk:1:1"]
      responses:
        '200':
          description: Pane statuses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaneStatusListResponse'
              example:
                statuses:
                  "hsk:1:0":
                    pane_status: "in_progress"
                  "hsk:1:1":
                    pane_status: "idle"
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags: [Panes]
      summary: Update pane status
      description: Set pane status to idle, in_progress, or done
      operationId: updatePaneStatus
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePaneStatusRequest'
            example:
              pane_key: "hsk:1:0"
              pane_status: "done"
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  pane_key:
                    type: string
                  pane_status:
                    type: string
              example:
                success: true
                pane_key: "hsk:1:0"
                pane_status: "done"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================================================
  # Tasks & Segments
  # ============================================================================
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: |
        HttpOnly session cookie set by /api/auth/login.
        Cookie settings: Max-Age=30 days, SameSite=Lax, Secure in production.
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "unauthorized"
            message: "Authentication required"
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "bad_request"
            message: "Invalid input"
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error type identifier
        message:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Optional error code

    # --------------------------------------------------------------------------
    # Auth
    # --------------------------------------------------------------------------
...
    Profile:
      type: object
      properties:
        id:
          type: integer
        profile_key:
          type: string
          description: Unique profile identifier (UUID or slug)
        name:
          type: string
        ctime:
          type: integer
          description: Creation timestamp
        mtime:
          type: integer
          description: Modification timestamp
    ProfileListResponse:
      type: object
      properties:
        profiles:
          type: array
          items:
            $ref: '#/components/schemas/Profile'
    CreateProfileRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        profile_key:
          type: string
          description: Optional custom key (auto-generated if not provided)
...
    Group:
      type: object
      properties:
        id:
          type: integer
        group_name:
          type: string
        sort_order:
          type: integer
        session_count:
          type: integer
          description: Number of sessions in this group
        ctime:
          type: integer
        mtime:
          type: integer
    GroupListResponse:
      type: object
      properties:
        groups:
          type: array
          items:
            $ref: '#/components/schemas/Group'
    CreateGroupRequest:
      type: object
      required:
        - profile_key
        - group_name
      properties:
        profile_key:
          type: string
        group_name:
          type: string
        sort_order:
          type: integer
          default: 0
...
    PaneStatusListResponse:
      type: object
      properties:
        statuses:
          type: object
          additionalProperties:
            type: object
            properties:
              pane_status:
                type: string
                enum: [idle, in_progress, done]
    UpdatePaneStatusRequest:
      type: object
      required:
        - pane_key
        - pane_status
      properties:
        pane_key:
          type: string
          description: "Format: sessionName:windowIndex:paneIndex"
        pane_status:
          type: string
          enum: [idle, in_progress, done]

    # --------------------------------------------------------------------------
    # Tasks
    # --------------------------------------------------------------------------
...